move and rename HtApi api layer object to smaller/discrete api layer objects like UserApi to be more descriptive
