+++ client/src/store/userSlice/index.ts
@@ -1,4 +1,6 @@
 import userReducer, { login, selectUserName, selectUser, selectUserState } from './user.slice';
+import { HaztrakUser } from './user.slice';
 
 export default userReducer;
 export { login, selectUserName, selectUser, selectUserState };
+export type { HaztrakUser };
